+++ packages/client/src/types.ts
@@ -222,12 +222,6 @@ export interface PolkadotClient {
   blocks$: Observable<BlockInfo>
 
   /**
-   * Observable of all the pinned blocks available. These are the blocks you can
-   * perform operations against.
-   */
-  pinnedBlocks$: Observable<Record<string, BlockInfo>>
-
-  /**
    * Observable to watch Block Body.
    *
    * @param hash  It can be a block hash, `"finalized"`, or `"best"`